Initial commit for revamping integration tests
