+++ docker_test.go
@@ -31,7 +31,6 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/gofrs/uuid"
 	redigo "github.com/gomodule/redigo/redis"
 	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
@@ -46,10 +45,7 @@ import (
 	"github.com/rudderlabs/rudder-server/services/streammanager/kafka/client/testutil"
 	"github.com/rudderlabs/rudder-server/testhelper"
 	"github.com/rudderlabs/rudder-server/testhelper/destination"
-	wht "github.com/rudderlabs/rudder-server/testhelper/warehouse"
 	"github.com/rudderlabs/rudder-server/utils/logger"
-	bq "github.com/rudderlabs/rudder-server/warehouse/bigquery"
-	"github.com/rudderlabs/rudder-server/warehouse/client"
 )
 
 var (
@@ -71,7 +67,6 @@ var (
 	MINIOContainer               *destination.MINIOResource
 	EventID                      string
 	VersionID                    string
-	runBigQueryTest              bool
 )
 
 type WebhookRecorder struct {
@@ -288,7 +283,6 @@ func TestMain(m *testing.M) {
 
 	flag.BoolVar(&hold, "hold", false, "hold environment clean-up after test execution until Ctrl+C is provided")
 	flag.BoolVar(&runIntegration, "integration", false, "run integration level tests")
-	flag.BoolVar(&runBigQueryTest, "bigqueryintegration", false, "run big query test")
 	flag.Parse()
 
 	if !runIntegration {
@@ -375,16 +369,6 @@ func run(m *testing.M) (int, error) {
 	_ = os.Setenv("WAREHOUSE_JOBS_DB_PORT", PostgresContainer.Port)
 	_ = os.Setenv("DEST_TRANSFORM_URL", TransformerContainer.TransformURL)
 
-	wht.InitWHConfig()
-
-	defer wht.SetWHPostgresDestination(pool)()
-	defer wht.SetWHClickHouseDestination(pool)()
-	defer wht.SetWHClickHouseClusterDestination(pool)()
-	defer wht.SetWHMssqlDestination(pool)()
-	defer wht.SetWHBigQueryDestination()()
-
-	AddWHSpecificSqlFunctionsToJobsDb()
-
 	httpPortInt, err := freeport.GetFreePort()
 	if err != nil {
 		return 0, fmt.Errorf("could not get free port for http: %w", err)
@@ -410,32 +394,17 @@ func run(m *testing.M) (int, error) {
 	writeKey = randString(27)
 	workspaceID = randString(27)
 	mapWorkspaceConfig := map[string]string{
-		"webhookUrl":                          webhookURL,
-		"disableDestinationwebhookUrl":        disableDestinationWebhookURL,
-		"writeKey":                            writeKey,
-		"workspaceId":                         workspaceID,
-		"postgresPort":                        PostgresContainer.Port,
-		"address":                             RedisContainer.RedisAddress,
-		"minioEndpoint":                       MINIOContainer.MinioEndpoint,
-		"minioBucketName":                     MINIOContainer.MinioBucketName,
-		"postgresEventWriteKey":               wht.Test.PGTest.WriteKey,
-		"clickHouseEventWriteKey":             wht.Test.CHTest.WriteKey,
-		"clickHouseClusterEventWriteKey":      wht.Test.CHClusterTest.WriteKey,
-		"mssqlEventWriteKey":                  wht.Test.MSSQLTest.WriteKey,
-		"rwhPostgresDestinationPort":          wht.Test.PGTest.Credentials.Port,
-		"rwhClickHouseDestinationPort":        wht.Test.CHTest.Credentials.Port,
-		"rwhClickHouseClusterDestinationPort": wht.Test.CHClusterTest.GetResource().Credentials.Port,
-		"rwhMSSqlDestinationPort":             wht.Test.MSSQLTest.Credentials.Port,
+		"webhookUrl":                   webhookURL,
+		"disableDestinationwebhookUrl": disableDestinationWebhookURL,
+		"writeKey":                     writeKey,
+		"workspaceId":                  workspaceID,
+		"postgresPort":                 PostgresContainer.Port,
+		"address":                      RedisContainer.RedisAddress,
+		"minioEndpoint":                MINIOContainer.MinioEndpoint,
+		"minioBucketName":              MINIOContainer.MinioBucketName,
 	}
 	if runtime.GOARCH != "arm64" || overrideArm64Check {
 		mapWorkspaceConfig["kafkaPort"] = KafkaContainer.Port
-	}
-	if runBigQueryTest && wht.Test.BQTest != nil {
-		mapWorkspaceConfig["bqEventWriteKey"] = wht.Test.BQTest.WriteKey
-		mapWorkspaceConfig["rwhBQProject"] = wht.Test.BQTest.Credentials.ProjectID
-		mapWorkspaceConfig["rwhBQLocation"] = wht.Test.BQTest.Credentials.Location
-		mapWorkspaceConfig["rwhBQBucketName"] = wht.Test.BQTest.Credentials.Bucket
-		mapWorkspaceConfig["rwhBQCredentials"] = wht.Test.BQTest.Credentials.CredentialsEscaped
 	}
 	workspaceConfigPath := createWorkspaceConfig(
 		"testdata/workspaceConfigTemplate.json",
@@ -914,720 +883,6 @@ func TestBeaconBatch(_ *testing.T) {
 		]
 	}`)
 	SendEvent(payload, "beacon/v1/batch", writeKey)
-}
-
-// Adding necessary sql functions which needs to be used during warehouse testing
-func AddWHSpecificSqlFunctionsToJobsDb() {
-	var err error
-	_, err = db.Exec(wht.Test.GatewayJobsSqlFunction)
-	if err != nil {
-		panic(fmt.Errorf("error occurred with executing gw jobs function for events count with err %v", err.Error()))
-	}
-
-	_, err = db.Exec(wht.Test.BatchRouterJobsSqlFunction)
-	if err != nil {
-		panic(fmt.Errorf("error occurred with executing brt jobs function for events count with err %s", err.Error()))
-	}
-}
-
-//Verify Event in WareHouse Postgres
-func TestWHPostgresDestination(t *testing.T) {
-	pgTest := wht.Test.PGTest
-
-	whDestTest := &wht.WareHouseDestinationTest{
-		Client: &client.Client{
-			SQL:  pgTest.DB,
-			Type: client.SQLClient,
-		},
-		EventsCountMap:     pgTest.EventsMap,
-		WriteKey:           pgTest.WriteKey,
-		UserId:             "userId_postgres",
-		Schema:             "postgres_wh_integration",
-		TableTestQueryFreq: pgTest.TableTestQueryFreq,
-	}
-	sendWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-
-	whDestTest.UserId = "userId_postgres_1"
-	sendUpdatedWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-}
-
-//Verify Event in WareHouse ClickHouse
-func TestWHClickHouseDestination(t *testing.T) {
-	chTest := wht.Test.CHTest
-
-	whDestTest := &wht.WareHouseDestinationTest{
-		Client: &client.Client{
-			SQL:  chTest.DB,
-			Type: client.SQLClient,
-		},
-		EventsCountMap:     chTest.EventsMap,
-		WriteKey:           chTest.WriteKey,
-		UserId:             "userId_clickhouse",
-		Schema:             "rudderdb",
-		TableTestQueryFreq: chTest.TableTestQueryFreq,
-	}
-	sendWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-
-	whDestTest.UserId = "userId_clickhouse_1"
-	sendUpdatedWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-}
-
-//Verify Event in WareHouse ClickHouse Cluster
-func TestWHClickHouseClusterDestination(t *testing.T) {
-	chClusterTest := wht.Test.CHClusterTest
-
-	whDestTest := &wht.WareHouseDestinationTest{
-		Client: &client.Client{
-			SQL:  chClusterTest.GetResource().DB,
-			Type: client.SQLClient,
-		},
-		EventsCountMap:     chClusterTest.EventsMap,
-		WriteKey:           chClusterTest.WriteKey,
-		UserId:             "userId_clickhouse_cluster",
-		Schema:             "rudderdb",
-		TableTestQueryFreq: chClusterTest.TableTestQueryFreq,
-	}
-	sendWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-
-	initWHClickHouseClusterModeSetup(t)
-
-	whDestTest.UserId = "userId_clickhouse_cluster_1"
-	sendUpdatedWHEvents(whDestTest)
-
-	// Update events count Map
-	// This is required as because of the cluster mode setup and distributed view, events are getting duplicated.
-	whDestTest.EventsCountMap = wht.EventsCountMap{
-		"identifies":    2,
-		"users":         2,
-		"tracks":        2,
-		"product_track": 2,
-		"pages":         2,
-		"screens":       2,
-		"aliases":       2,
-		"groups":        2,
-		"gateway":       6,
-		"batchRT":       8,
-	}
-	whDestinationTest(t, whDestTest)
-}
-
-func TestWHBigQuery(t *testing.T) {
-	if runBigQueryTest == false {
-		t.Skip("Big query integration skipped. use -bigqueryintegration to add this test ")
-
-	}
-	if wht.Test.BQTest == nil {
-		fmt.Println("Error in ENV variable BIGQUERY_INTEGRATION_TEST_USER_CRED")
-		t.FailNow()
-	}
-	//Disabling big query dedup
-	config.SetBool("Warehouse.bigquery.isDedupEnabled", false)
-	bq.Init()
-	bqTest := wht.Test.BQTest
-	randomness := strings.ReplaceAll(uuid.Must(uuid.NewV4()).String(), "-", "")
-
-	whDestTest := &wht.WareHouseDestinationTest{
-		Client: &client.Client{
-			BQ:   bqTest.DB,
-			Type: client.BQClient,
-		},
-		EventsCountMap:     bqTest.EventsMap,
-		WriteKey:           bqTest.WriteKey,
-		UserId:             fmt.Sprintf("userId_bq_%s", randomness),
-		Schema:             "rudderstack_sample_http_source",
-		BQContext:          bqTest.Context,
-		Tables:             bqTest.Tables,
-		PrimaryKeys:        bqTest.PrimaryKeys,
-		TableTestQueryFreq: bqTest.TableTestQueryFreq,
-	}
-
-	whDestTest.MessageId = uuid.Must(uuid.NewV4()).String()
-
-	whDestTest.EventsCountMap = wht.EventsCountMap{
-		"identifies": 2,
-		"tracks":     2,
-		"pages":      2,
-		"screens":    2,
-		"aliases":    2,
-		"groups":     2,
-	}
-	sendWHEvents(whDestTest)
-
-	whDestTest.EventsCountMap = wht.EventsCountMap{
-		"identifies":    2,
-		"users":         1,
-		"tracks":        2,
-		"product_track": 2,
-		"pages":         2,
-		"screens":       2,
-		"aliases":       2,
-		"_groups":       2,
-		"gateway":       12,
-		"batchRT":       16,
-	}
-
-	whDestinationTest(t, whDestTest)
-	//Enabling big query dedup
-	config.SetBool("Warehouse.bigquery.isDedupEnabled", true)
-	bq.Init()
-
-	whDestTest.EventsCountMap = wht.EventsCountMap{
-		"identifies": 2,
-		"tracks":     2,
-		"pages":      2,
-		"screens":    2,
-		"aliases":    2,
-		"groups":     2,
-	}
-
-	sendWHEvents(whDestTest)
-
-	whDestTest.EventsCountMap = wht.EventsCountMap{
-		"identifies":    2,
-		"users":         1,
-		"tracks":        2,
-		"product_track": 2,
-		"pages":         2,
-		"screens":       2,
-		"aliases":       2,
-		"_groups":       2,
-		"gateway":       24,
-		"batchRT":       32,
-	}
-
-	whDestinationTest(t, whDestTest)
-
-}
-
-//Verify Event in WareHouse MSSQL
-func TestWHMssqlDestination(t *testing.T) {
-	MssqlTest := wht.Test.MSSQLTest
-
-	whDestTest := &wht.WareHouseDestinationTest{
-		Client: &client.Client{
-			SQL:  MssqlTest.DB,
-			Type: client.SQLClient,
-		},
-		EventsCountMap:     MssqlTest.EventsMap,
-		WriteKey:           MssqlTest.WriteKey,
-		UserId:             "userId_mssql",
-		Schema:             "mssql_wh_integration",
-		TableTestQueryFreq: MssqlTest.TableTestQueryFreq,
-	}
-	sendWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-
-	whDestTest.UserId = "userId_mssql_1"
-	sendUpdatedWHEvents(whDestTest)
-	whDestinationTest(t, whDestTest)
-}
-
-func getMessageId(wdt *wht.WareHouseDestinationTest) string {
-	if wdt.MessageId == "" {
-		return uuid.Must(uuid.NewV4()).String()
-	}
-	return wdt.MessageId
-}
-
-// sendWHEvents Sending warehouse events
-func sendWHEvents(wdt *wht.WareHouseDestinationTest) {
-	// Sending identify event
-	if identify, exists := wdt.EventsCountMap["identifies"]; exists {
-		for i := 0; i < identify; i++ {
-			payloadIdentify := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "identify",
-			"eventOrderNo":"1",
-			"context": {
-			  "traits": {
-				 "trait1": "new-val"
-			  }
-			},
-			"timestamp": "2020-02-02T00:23:09.544Z"
-		  }`, wdt.UserId, getMessageId(wdt)))
-			SendEvent(payloadIdentify, "identify", wdt.WriteKey)
-		}
-	}
-
-	// Sending track event
-	if track, exists := wdt.EventsCountMap["tracks"]; exists {
-		for i := 0; i < track; i++ {
-			payloadTrack := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "track",
-			"event": "Product Track",
-			"properties": {
-			  "review_id": "12345",
-			  "product_id" : "123",
-			  "rating" : 3.0,
-			  "review_body" : "Average product, expected much more."
-			}
-		  }`, wdt.UserId, getMessageId(wdt)))
-			SendEvent(payloadTrack, "track", wdt.WriteKey)
-		}
-	}
-
-	// Sending page event
-	if page, exists := wdt.EventsCountMap["pages"]; exists {
-		for i := 0; i < page; i++ {
-			payloadPage := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "page",
-			"name": "Home",
-			"properties": {
-			  "title": "Home | RudderStack",
-			  "url": "http://www.rudderstack.com"
-			}
-		  }`, wdt.UserId, getMessageId(wdt)))
-			SendEvent(payloadPage, "page", wdt.WriteKey)
-		}
-	}
-
-	// Sending screen event
-	if screen, exists := wdt.EventsCountMap["screens"]; exists {
-		for i := 0; i < screen; i++ {
-			payloadScreen := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "screen",
-			"name": "Main",
-			"properties": {
-			  "prop_key": "prop_value"
-			}
-		  }`, wdt.UserId, getMessageId(wdt)))
-			SendEvent(payloadScreen, "screen", wdt.WriteKey)
-		}
-	}
-
-	// Sending alias event
-	if alias, exists := wdt.EventsCountMap["aliases"]; exists {
-		for i := 0; i < alias; i++ {
-			payloadAlias := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "alias",
-			"previousId": "name@surname.com"
-		  }`, wdt.UserId, getMessageId(wdt)))
-			SendEvent(payloadAlias, "alias", wdt.WriteKey)
-		}
-	}
-
-	// Sending group event
-	if group, exists := wdt.EventsCountMap["groups"]; exists {
-		for i := 0; i < group; i++ {
-			payloadGroup := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "group",
-			"groupId": "groupId",
-			"traits": {
-			  "name": "MyGroup",
-			  "industry": "IT",
-			  "employees": 450,
-			  "plan": "basic"
-			}
-		  }`, wdt.UserId, getMessageId(wdt)))
-			SendEvent(payloadGroup, "group", wdt.WriteKey)
-		}
-	}
-}
-
-// sendUpdatedWHEvents Sending updated warehouse events
-func sendUpdatedWHEvents(wdt *wht.WareHouseDestinationTest) {
-	// Sending identify event
-	if identify, exists := wdt.EventsCountMap["identifies"]; exists {
-		for i := 0; i < identify; i++ {
-			payloadIdentify := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "identify",
-			"context": {
-			  "traits": {
-				 "trait1": "new-val"
-			  },
-			  "ip": "14.5.67.21",
-			  "library": {
-				  "name": "http"
-			  }
-			},
-			"timestamp": "2020-02-02T00:23:09.544Z"
-		  }`, wdt.UserId, uuid.Must(uuid.NewV4()).String()))
-			SendEvent(payloadIdentify, "identify", wdt.WriteKey)
-		}
-	}
-
-	// Sending track event
-	if track, exists := wdt.EventsCountMap["tracks"]; exists {
-		for i := 0; i < track; i++ {
-			payloadTrack := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "track",
-			"event": "Product Track",
-			"properties": {
-			  "review_id": "12345",
-			  "product_id" : "123",
-			  "rating" : 3.0,
-			  "revenue" : 4.99,
-			  "review_body" : "Average product, expected much more."
-			},
-			"context": {
-				"ip": "14.5.67.21",
-				"library": {
-					"name": "http"
-				}
-			}
-		  }`, wdt.UserId, uuid.Must(uuid.NewV4()).String()))
-			SendEvent(payloadTrack, "track", wdt.WriteKey)
-		}
-	}
-
-	// Sending page event
-	if page, exists := wdt.EventsCountMap["pages"]; exists {
-		for i := 0; i < page; i++ {
-			payloadPage := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "page",
-			"name": "Home",
-			"properties": {
-			  "title": "Home | RudderStack",
-			  "url": "http://www.rudderstack.com"
-			},
-			"context": {
-				"ip": "14.5.67.21",
-				"library": {
-					"name": "http"
-				}
-			  }
-		  }`, wdt.UserId, uuid.Must(uuid.NewV4()).String()))
-			SendEvent(payloadPage, "page", wdt.WriteKey)
-		}
-	}
-
-	// Sending screen event
-	if screen, exists := wdt.EventsCountMap["screens"]; exists {
-		for i := 0; i < screen; i++ {
-			payloadScreen := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "screen",
-			"name": "Main",
-			"properties": {
-			  "prop_key": "prop_value"
-			},
-			"context": {
-				"ip": "14.5.67.21",
-				"library": {
-					"name": "http"
-				}
-			  }
-		  }`, wdt.UserId, uuid.Must(uuid.NewV4()).String()))
-			SendEvent(payloadScreen, "screen", wdt.WriteKey)
-		}
-	}
-
-	// Sending alias event
-	if alias, exists := wdt.EventsCountMap["aliases"]; exists {
-		for i := 0; i < alias; i++ {
-			payloadAlias := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "alias",
-			"previousId": "name@surname.com",
-			"context": {
-				"ip": "14.5.67.21",
-				"library": {
-					"name": "http"
-				}
-            }
-		  }`, wdt.UserId, uuid.Must(uuid.NewV4()).String()))
-			SendEvent(payloadAlias, "alias", wdt.WriteKey)
-		}
-	}
-
-	// Sending group event
-	if group, exists := wdt.EventsCountMap["groups"]; exists {
-		for i := 0; i < group; i++ {
-			payloadGroup := strings.NewReader(fmt.Sprintf(`{
-			"userId": "%s",
-			"messageId":"%s",
-			"type": "group",
-			"groupId": "groupId",
-			"traits": {
-			  "name": "MyGroup",
-			  "industry": "IT",
-			  "employees": 450,
-			  "plan": "basic"
-			},
-			"context": {
-				"ip": "14.5.67.21",
-				"library": {
-					"name": "http"
-				}
-			}
-		  }`, wdt.UserId, uuid.Must(uuid.NewV4()).String()))
-			SendEvent(payloadGroup, "group", wdt.WriteKey)
-		}
-	}
-}
-
-// whDestinationTest Checking warehouse destination
-func whDestinationTest(t *testing.T, wdt *wht.WareHouseDestinationTest) {
-	whGatewayTest(t, wdt)
-	whBatchRouterTest(t, wdt)
-	whTablesTest(t, wdt)
-}
-
-// whGatewayTest Checking for gateway jobs
-func whGatewayTest(t *testing.T, wdt *wht.WareHouseDestinationTest) {
-	require.Contains(t, wdt.EventsCountMap, "gateway")
-	gwEvents := wdt.EventsCountMap["gateway"]
-	// Checking for the gateway jobs
-	t.Log("Checking for the gateway jobs")
-	require.Eventually(t, func() bool {
-		var count int64
-		jobSqlStatement := fmt.Sprintf(`select count(*) from gw_jobs_for_user_id_and_write_key('%s', '%s') as job_ids`, wdt.UserId, wdt.WriteKey)
-		err := db.QueryRow(jobSqlStatement).Scan(&count)
-		require.Equal(t, err, nil)
-		return count == int64(gwEvents)
-	}, time.Minute, 10*time.Millisecond)
-
-	// Getting gateway job ids
-	t.Log("Getting gateway job ids")
-	var jobIds []string
-	require.Eventually(t, func() bool {
-		jobIds = make([]string, 0)
-		jobSqlStatement := fmt.Sprintf(`select * from gw_jobs_for_user_id_and_write_key('%s', '%s') as job_ids`, wdt.UserId, wdt.WriteKey)
-		rows, err := db.Query(jobSqlStatement)
-		require.Equal(t, err, nil)
-
-		defer func() { _ = rows.Close() }()
-		for rows.Next() {
-			var jobId int64
-			err = rows.Scan(&jobId)
-			require.Equal(t, err, nil)
-			jobIds = append(jobIds, fmt.Sprint(jobId))
-		}
-		return gwEvents == len(jobIds)
-	}, time.Minute, 10*time.Millisecond)
-
-	// Checking for the gateway jobs state
-	t.Log("Checking for the gateway jobs state")
-	require.Eventually(t, func() bool {
-		var count int64
-		jobsSqlStatement := fmt.Sprintf("select count(*) from gw_job_status_1 where job_id in (%s) and job_state = 'succeeded'", strings.Join(jobIds, ","))
-		err := db.QueryRow(jobsSqlStatement).Scan(&count)
-		require.Equal(t, nil, err)
-		return count == int64(gwEvents)
-	}, time.Minute, 10*time.Millisecond)
-}
-
-// whBatchRouterTest Checking for batch router jobs
-func whBatchRouterTest(t *testing.T, wdt *wht.WareHouseDestinationTest) {
-	require.Contains(t, wdt.EventsCountMap, "batchRT")
-	brtEvents := wdt.EventsCountMap["batchRT"]
-
-	// Checking for the batch router jobs
-	t.Log("Checking for the batch router jobs")
-	require.Eventually(t, func() bool {
-		var count int64
-		jobsSqlStatement := fmt.Sprintf(`select count(*) from brt_jobs_for_user_id('%s') as job_ids`, wdt.UserId)
-		err := db.QueryRow(jobsSqlStatement).Scan(&count)
-		require.Equal(t, err, nil)
-		return count == int64(brtEvents)
-	}, 2*time.Minute, 100*time.Millisecond)
-
-	// Getting batch router job ids
-	t.Log("Getting batch router job ids")
-	var jobIds []string
-	require.Eventually(t, func() bool {
-		jobIds = make([]string, 0)
-		jobSqlStatement := fmt.Sprintf(`select * from brt_jobs_for_user_id('%s') as job_ids`, wdt.UserId)
-		rows, err := db.Query(jobSqlStatement)
-		require.Equal(t, err, nil)
-
-		defer func() { _ = rows.Close() }()
-		for rows.Next() {
-			var jobId int64
-			err = rows.Scan(&jobId)
-			require.Equal(t, err, nil)
-			jobIds = append(jobIds, fmt.Sprint(jobId))
-		}
-		return brtEvents == len(jobIds)
-	}, 2*time.Minute, 100*time.Millisecond)
-
-	// Checking for the batch router jobs state
-	t.Log("Checking for the batch router jobs state")
-	require.Eventually(t, func() bool {
-		var count int64
-		jobsSqlStatement := fmt.Sprintf("select count(*) from batch_rt_job_status_1 where job_id in (%s) and job_state = 'succeeded'", strings.Join(jobIds, ","))
-		err := db.QueryRow(jobsSqlStatement).Scan(&count)
-		require.Equal(t, err, nil)
-		return count == int64(brtEvents)
-	}, 2*time.Minute, 100*time.Millisecond)
-}
-
-func getQueryCount(cl *client.Client, statement string) (int64, error) {
-	result, err := cl.Query(statement, client.Read)
-	if err != nil {
-		return 0, err
-	}
-	return strconv.ParseInt(result.Values[0][0], 10, 64)
-}
-
-// whTablesTest Checking warehouse
-func whTablesTest(t *testing.T, wdt *wht.WareHouseDestinationTest) {
-	tables := []string{"identifies", "users", "tracks", "product_track", "pages", "screens", "aliases", "groups"}
-	primaryKeys := []string{"user_id", "id", "user_id", "user_id", "user_id", "user_id", "user_id", "user_id"}
-
-	if len(wdt.Tables) != 0 {
-		tables = wdt.Tables
-	}
-	if len(wdt.PrimaryKeys) != 0 {
-		primaryKeys = wdt.PrimaryKeys
-	}
-
-	for idx, table := range tables {
-		require.Contains(t, wdt.EventsCountMap, table)
-		tableCount := wdt.EventsCountMap[table]
-		require.Eventually(t, func() bool {
-			var count int64
-			sqlStatement := fmt.Sprintf("select count(*) from %s.%s where %s = '%s'", wdt.Schema, table, primaryKeys[idx], wdt.UserId)
-			count, _ = getQueryCount(wdt.Client, sqlStatement)
-			return count == int64(tableCount)
-		}, 2*time.Minute, wdt.TableTestQueryFreq)
-	}
-}
-
-// initWHClickHouseClusterModeSetup Initialize cluster mode setup
-func initWHClickHouseClusterModeSetup(t *testing.T) {
-	type ColumnInfoT struct {
-		ColumnName string
-		ColumnType string
-	}
-
-	chClusterTest := wht.Test.CHClusterTest
-	tables := []string{"identifies", "users", "tracks", "product_track", "pages", "screens", "aliases", "groups"}
-	tableColumnInfoMap := map[string][]ColumnInfoT{
-		"identifies": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-		"product_track": {
-			{
-				ColumnName: "revenue",
-				ColumnType: "Nullable(Float64)",
-			},
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-		"tracks": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-		"users": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "SimpleAggregateFunction(anyLast, Nullable(String))",
-			},
-		},
-		"pages": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-		"screens": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-		"aliases": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-		"groups": {
-			{
-				ColumnName: "context_passed_ip",
-				ColumnType: "Nullable(String)",
-			},
-			{
-				ColumnName: "context_library_name",
-				ColumnType: "Nullable(String)",
-			},
-		},
-	}
-
-	// Rename tables to tables_shard
-	for _, table := range tables {
-		sqlStatement := fmt.Sprintf("RENAME TABLE %[1]s to %[1]s_shard ON CLUSTER rudder_cluster;", table)
-		_, err := chClusterTest.GetResource().DB.Exec(sqlStatement)
-		require.Equal(t, err, nil)
-	}
-
-	// Create distribution views for tables
-	for _, table := range tables {
-		sqlStatement := fmt.Sprintf("CREATE TABLE rudderdb.%[1]s ON CLUSTER 'rudder_cluster' AS rudderdb.%[1]s_shard ENGINE = Distributed('rudder_cluster', rudderdb, %[1]s_shard, cityHash64(concat(toString(received_at), id)));", table)
-		_, err := chClusterTest.GetResource().DB.Exec(sqlStatement)
-		require.Equal(t, err, nil)
-	}
-
-	// Alter columns to all the cluster tables
-	for _, chResource := range chClusterTest.Resources {
-		for tableName, columnInfos := range tableColumnInfoMap {
-			for _, columnInfo := range columnInfos {
-				sqlStatement := fmt.Sprintf("ALTER TABLE rudderdb.%[1]s_shard ADD COLUMN IF NOT EXISTS %[2]s %[3]s;", tableName, columnInfo.ColumnName, columnInfo.ColumnType)
-				_, err := chResource.DB.Exec(sqlStatement)
-				require.Equal(t, err, nil)
-			}
-		}
-	}
 }
 
 type testLogger struct{ logger.LoggerI }
