+++ warehouse/integration__tests/testdata/clickhouse/cluster/clickhouse01/config.xml
