+++ warehouse/integration__tests/testdata/clickhouse/cluster/clickhouse02/users.xml
