+++ warehouse/integration__tests/testdata/clickhouse/cluster/clickhouse03/config.xml
