+++ warehouse/integration__tests/testhelper/helper.go
@@ -1,0 +1,104 @@
+package testhelper
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"log"
+	"math/rand"
+	"net/http"
+	"strings"
+	"time"
+)
+
+func WaitUntilReady(ctx context.Context, endpoint string, atMost, interval time.Duration, caller string) {
+	probe := time.NewTicker(interval)
+	timeout := time.After(atMost)
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-timeout:
+			log.Panicf("application was not ready after %s, for the end point: %s, caller: %s", atMost, endpoint, caller)
+		case <-probe.C:
+			resp, err := http.Get(endpoint)
+			if err != nil {
+				continue
+			}
+			if resp.StatusCode == http.StatusOK {
+				log.Println("application ready")
+				return
+			}
+		}
+	}
+}
+
+func RandString(n int) string {
+	var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
+
+	s := make([]rune, n)
+	for i := range s {
+		s[i] = letters[rand.Intn(len(letters))]
+	}
+	return string(s)
+}
+
+func JsonEscape(i string) (string, error) {
+	b, err := json.Marshal(i)
+	if err != nil {
+		return "", fmt.Errorf("could not escape big query JSON credentials for workspace config with error: %s", err.Error())
+	}
+	return strings.Trim(string(b), `"`), nil
+}
+
+func GatewayJobsSqlFunction() string {
+	return `CREATE OR REPLACE FUNCTION gw_jobs_for_user_id_and_write_key(user_id varchar, write_key varchar)
+								RETURNS TABLE
+										(
+											job_id varchar
+										)
+							AS
+							$$
+							DECLARE
+								table_record RECORD;
+								batch_record jsonb;
+							BEGIN
+								FOR table_record IN SELECT * FROM gw_jobs_1 where (event_payload ->> 'writeKey') = write_key
+									LOOP
+										FOR batch_record IN SELECT * FROM jsonb_array_elements((table_record.event_payload ->> 'batch')::jsonb)
+											LOOP
+												if batch_record ->> 'userId' != user_id THEN
+													CONTINUE;
+												END IF;
+												job_id := table_record.job_id;
+												RETURN NEXT;
+												EXIT;
+											END LOOP;
+									END LOOP;
+							END;
+							$$ LANGUAGE plpgsql`
+}
+
+func BatchRouterJobsSqlFunction() string {
+	return `CREATE OR REPLACE FUNCTION brt_jobs_for_user_id(user_id varchar)
+									RETURNS TABLE
+											(
+												job_id varchar
+											)
+								AS
+								$$
+								DECLARE
+									table_record  RECORD;
+									event_payload jsonb;
+								BEGIN
+									FOR table_record IN SELECT * FROM batch_rt_jobs_1
+										LOOP
+											event_payload = (table_record.event_payload ->> 'data')::jsonb;
+											if event_payload ->> 'user_id' = user_id Or event_payload ->> 'id' = user_id THEN
+												job_id := table_record.job_id;
+												RETURN NEXT;
+											END IF;
+										END LOOP;
+								END ;
+								$$ LANGUAGE plpgsql`
+}
